% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-content-type.R
\name{get_content_type}
\alias{get_content_type}
\title{Discover MIME type of a file based on contents}
\usage{
get_content_type(path)
}
\arguments{
\item{path}{path to a file}
}
\description{
Discover MIME type of a file based on contents
}
\details{
Initial in-R header mapping logic borrowed from \code{MimeTypes.java} from
\href{https://github.com/Servoy/servoy-client}{servoy-client}
}
