% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplemagic-package.R
\docType{package}
\name{simplemagic}
\alias{simplemagic}
\alias{simplemagic-package}
\title{Lightweight File 'MIME' Type Detection Based On Contents or Extension}
\description{
Lightweight File 'MIME' Type Detection Based On Contents or Extension
}
\author{
Bob Rudis (bob@rud.is)
}
